analyze user turns to find out act combinations & requested intent/slots combinations
