--- data/crosswoz/processed_demo_2303.json
@@ -1,212 +1,0 @@
-[
-  {
-    "UserTurns": 96,
-    "SysTurns": 97,
-    "Slots": [
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "人均消费",
-        "Values": {
-          "single": "50-100元"
-        },
-        "Filled": false
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "推荐菜",
-        "Values": {
-          "multi": [
-            "美食街"
-          ]
-        },
-        "Filled": false
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "名称",
-        "Values": {},
-        "Filled": false
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "营业时间",
-        "Values": {},
-        "Filled": false
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "周边景点",
-        "Values": {},
-        "Filled": false
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "名称",
-        "Values": {
-          "single": "出现在id=1的周边景点里"
-        },
-        "Filled": false
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "评分",
-        "Values": {
-          "single": "4.5分以上"
-        },
-        "Filled": false
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "地址",
-        "Values": {},
-        "Filled": false
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "电话",
-        "Values": {},
-        "Filled": false
-      },
-      {
-        "ID": 3,
-        "Group": "酒店",
-        "Name": "名称",
-        "Values": {
-          "single": "桔子水晶酒店(北京安贞店)"
-        },
-        "Filled": false
-      },
-      {
-        "ID": 3,
-        "Group": "酒店",
-        "Name": "电话",
-        "Values": {},
-        "Filled": false
-      }
-    ],
-    "FilledSlots": [
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "人均消费",
-        "Values": {
-          "single": "50-100元"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "推荐菜",
-        "Values": {
-          "multi": [
-            "美食街"
-          ]
-        },
-        "Filled": true
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "名称",
-        "Values": {
-          "single": "鲜鱼口老字号美食街"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "营业时间",
-        "Values": {
-          "single": "周一至周日 10:00-22:00"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 1,
-        "Group": "餐馆",
-        "Name": "周边景点",
-        "Values": {
-          "multi": [
-            "天安门广场",
-            "前门大街",
-            "恭王府",
-            "故宫"
-          ]
-        },
-        "Filled": true
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "名称",
-        "Values": {
-          "single": "故宫"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "评分",
-        "Values": {
-          "single": "4.5分以上"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "地址",
-        "Values": {
-          "single": "北京市东城区景山前街4号"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 2,
-        "Group": "景点",
-        "Name": "电话",
-        "Values": {
-          "single": "010-85007938"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 3,
-        "Group": "酒店",
-        "Name": "名称",
-        "Values": {
-          "single": "桔子水晶酒店(北京安贞店)"
-        },
-        "Filled": true
-      },
-      {
-        "ID": 3,
-        "Group": "酒店",
-        "Name": "电话",
-        "Values": {
-          "single": "010-84273030"
-        },
-        "Filled": true
-      }
-    ],
-    "dialogue_id": "2303",
-    "task description": [
-      "你要去一个餐馆(id=1)用餐。你希望餐馆的人均消费是50-100元的。你想吃的菜肴是美食街。你想知道这个餐馆的名称、营业时间、周边景点。",
-      "你要去id=1附近的景点(id=2)游玩。你希望景点的评分是4.5分以上。你想知道这个景点的地址、电话。",
-      "你要去名叫桔子水晶酒店(北京安贞店)的酒店(id=3)住宿。你想知道这个酒店的电话。"
-    ],
-    "type": "不独立多领域"
-  }
-]
